requirements.txt
ultralytics
sahi
timm
opencv-python-headless
tqdm
numpy
Pillow
pandas